feat: HandlerAdapterRegistry 래핑클래스 추가
