+++ app/src/main/java/com/techcourse/DispatcherServlet.java
@@ -1,7 +1,5 @@
 package com.techcourse;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 
 import jakarta.servlet.ServletException;
@@ -17,7 +15,7 @@ import com.interface21.webmvc.servlet.View;
 import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerAdapter;
 import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerMapping;
 import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
+import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapterRegistry;
 import com.interface21.webmvc.servlet.mvc.tobe.HandlerMappingRegistry;
 
 public class DispatcherServlet extends HttpServlet {
@@ -25,30 +23,21 @@ public class DispatcherServlet extends HttpServlet {
     private static final long serialVersionUID = 1L;
     private static final Logger log = LoggerFactory.getLogger(DispatcherServlet.class);
 
-    private HandlerMappingRegistry handlerMappingRegistry;
-    private List<HandlerAdapter> handlerAdapters;
+    private final HandlerMappingRegistry handlerMappingRegistry;
+    private final HandlerAdapterRegistry handlerAdapterRegistry;
 
     public DispatcherServlet() {
         handlerMappingRegistry = new HandlerMappingRegistry();
-    }
-
-    public void addHandlerAdapter(HandlerAdapter handlerAdapter) {
-        handlerAdapters.add(handlerAdapter);
-    }
-
-    public void addHandlerMapping(HandlerMapping handlerMapping) {
-        handlerMappingRegistry.addHandlerMapping(handlerMapping);
+        handlerAdapterRegistry = new HandlerAdapterRegistry();
     }
 
     @Override
     public void init() {
         handlerMappingRegistry.addHandlerMapping(new ManualHandlerMapping());
-        handlerMappingRegistry.addHandlerMapping
-                (new AnnotationHandlerMapping(DispatcherServlet.class.getPackage().getName()));
-
-        handlerAdapters = new ArrayList<>();
-        handlerAdapters.add(new ManualHandlerAdapter());
-        handlerAdapters.add(new AnnotationHandlerAdapter());
+        handlerMappingRegistry.addHandlerMapping(
+                new AnnotationHandlerMapping(DispatcherServlet.class.getPackage().getName()));
+        handlerAdapterRegistry.addHandlerAdapter(new ManualHandlerAdapter());
+        handlerAdapterRegistry.addHandlerAdapter(new AnnotationHandlerAdapter());
     }
 
     @Override
@@ -59,21 +48,13 @@ public class DispatcherServlet extends HttpServlet {
 
         try {
             Object handler = handlerMappingRegistry.getHandler(request);
-            HandlerAdapter handlerAdapter = getHandlerAdapter(handler, request);
+            HandlerAdapter handlerAdapter = handlerAdapterRegistry.getHandlerAdapter(handler);
             ModelAndView modelAndView = handlerAdapter.handle(request, response, handler);
             render(modelAndView, request, response);
         } catch (Throwable e) {
             log.error("Exception : {}", e.getMessage(), e);
             throw new ServletException(e.getMessage());
         }
-    }
-
-    private HandlerAdapter getHandlerAdapter(Object handler, HttpServletRequest request) throws ServletException {
-        return handlerAdapters.stream()
-                .filter(handlerAdapter -> handlerAdapter.support(handler))
-                .findFirst()
-                .orElseThrow(() ->
-                        new ServletException("No handlerAdapter found for request URI: " + request.getRequestURI()));
     }
 
     private void render(ModelAndView modelAndView, HttpServletRequest request, HttpServletResponse response)
