+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/HandlerAdapterRegistry.java
@@ -1,0 +1,26 @@
+package com.interface21.webmvc.servlet.mvc.tobe;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import jakarta.servlet.ServletException;
+
+public class HandlerAdapterRegistry {
+
+    private final List<HandlerAdapter> handlerAdapters;
+
+    public HandlerAdapterRegistry() {
+        handlerAdapters = new ArrayList<>();
+    }
+
+    public void addHandlerAdapter(HandlerAdapter handlerAdapter) {
+        handlerAdapters.add(handlerAdapter);
+    }
+
+    public HandlerAdapter getHandlerAdapter(Object handler) throws ServletException {
+        return handlerAdapters.stream()
+                .filter(handlerAdapter -> handlerAdapter.support(handler))
+                .findFirst()
+                .orElseThrow(() -> new ServletException("No handlerAdapter found"));
+    }
+}
